Created some composition polymorphism
